Replace lg.RequestLogger with our own simpler request logger
